jwt - isGuestMode validation added

+++ src/main/java/edu/ilyav/api/cotrollers/EducationController.java
@@ -12,29 +12,45 @@ import java.util.List;
 
 @RestController
 @RequestMapping("/rest/private/education")
-public class EducationController {
+public class EducationController extends BaseController {
 
 	@Autowired
 	private EducationService educationService;
 
 	@RequestMapping(value = "/add", method = RequestMethod.POST)
-	public Education add(@RequestBody Education education) throws ResourceNotFoundException {
-		return educationService.saveOrUpdate(education);
+	public Education add(HttpServletRequest request, @RequestBody Education education) throws ResourceNotFoundException {
+		if(isGuestMode(request))
+			throw new ResourceNotFoundException("You do not have this privilege in Guest mode.");
+		else {
+			return educationService.saveOrUpdate(education);
+		}
 	}
 
 	@RequestMapping(value = "/update", method = RequestMethod.PUT)
-	public Education update(@RequestBody Education education) throws ResourceNotFoundException {
-		return educationService.saveOrUpdate(education);
+	public Education update(HttpServletRequest request, @RequestBody Education education) throws ResourceNotFoundException {
+		if(isGuestMode(request))
+			throw new ResourceNotFoundException("You do not have this privilege in Guest mode.");
+		else {
+			return educationService.saveOrUpdate(education);
+		}
 	}
 
 	@RequestMapping(value = "/delete/{id}", method = RequestMethod.DELETE)
-	public WebResponse delete(@PathVariable Long id) throws Exception {
-		return educationService.delete(id);
+	public WebResponse delete(HttpServletRequest request, @PathVariable Long id) throws Exception {
+		if(isGuestMode(request))
+			throw new ResourceNotFoundException("You do not have this privilege in Guest mode.");
+		else {
+			return educationService.delete(id);
+		}
 	}
 
 	@RequestMapping(value = "/all", method = RequestMethod.GET)
-	public List<Education> getAll() {
-		return educationService.findAll();
+	public List<Education> getAll(HttpServletRequest request) throws ResourceNotFoundException {
+		if(isGuestMode(request))
+			throw new ResourceNotFoundException("You do not have this privilege in Guest mode.");
+		else {
+			return educationService.findAll();
+		}
 	}
