+++ src/main/java/edu/ilyav/api/cotrollers/RoleController.java
@@ -6,33 +6,46 @@ import edu.ilyav.api.service.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 
 @RestController
 @RequestMapping("/rest/private/role")
-public class RoleController {
+public class RoleController extends BaseController {
 
 	@Autowired
 	private RoleService roleService;
 
 	@RequestMapping(value = "/add", method = RequestMethod.POST)
-	public Role add(@RequestBody Role role) throws ResourceNotFoundException {
-		return roleService.saveOrUpdate(role);
+	public Role add(HttpServletRequest req, @RequestBody Role role) throws ResourceNotFoundException {
+		if(isGuestMode(req))
+			throw new ResourceNotFoundException("You do not have this privilege in Guest mode.");
+		else
+			return roleService.saveOrUpdate(role);
 	}
 
 	@RequestMapping(value = "/update", method = RequestMethod.PUT)
-	public Role update(@RequestBody Role role) throws ResourceNotFoundException {
-		return roleService.saveOrUpdate(role);
+	public Role update(HttpServletRequest req, @RequestBody Role role) throws ResourceNotFoundException {
+		if(isGuestMode(req))
+			throw new ResourceNotFoundException("You do not have this privilege in Guest mode.");
+		else
+			return roleService.saveOrUpdate(role);
 	}
 
 	@RequestMapping(value = "/delete/{id}", method = RequestMethod.DELETE)
-	public void delete(@PathVariable Long id) throws Exception {
-		roleService.delete(id);
+	public void delete(HttpServletRequest req, @PathVariable Long id) throws Exception {
+		if(isGuestMode(req))
+			throw new ResourceNotFoundException("You do not have this privilege in Guest mode.");
+		else
+			roleService.delete(id);
 	}
 
 	@RequestMapping(value = "/all", method = RequestMethod.GET)
-	public List<Role> getAll() {
-		return roleService.findAll();
+	public List<Role> getAll(HttpServletRequest req) throws ResourceNotFoundException {
+		if(isGuestMode(req))
+			throw new ResourceNotFoundException("You do not have this privilege in Guest mode.");
+		else
+			return roleService.findAll();
 	}
